+++ docs/changelog.md
@@ -10,9 +10,9 @@
 
 ## Версия 1.4.5
 - Теперь [mineTracks](/func?id=minetracks) может искать ключевые слова в названиях альбомов и самих треках. 
-- В `mineTracks` аргумент `playlistCount` переименован в `itemCount`.
+- В `mineTracks` аргумент `playlistCount` **переименован** в `itemCount`.
 - Новая функция у Filter: [replaceWithSimilar](/func?id=replacewithsimilar).
-- Новая функция у Lastfm: [getSimilarArtists](/func?id=getsimilarartists-1)
+- Новая функция у Lastfm: [getSimilarArtists](/func?id=getsimilarartists).
 
 ## Версия 1.4.4
 - Новый фильтр [removeUnavailable](/func?id=removeunavailable).